[nav-sim] Implement notifying when rover crashes into gate.
The gates are still note drawn completely to scale and it is possible
to appear to be driving through a gate post when you actually are not.
This, along with a more fleshed out logic for crashes will come in a
later commit but this functionality was needed now for testing even if the
UI was not perfect.