[nav] Connect indicator lights to programs
resolves #310

Implement components to monitor our connection to the nav,
localization, and perception programs by listening for LCM
messages from them. If we have not received a message in the
past 3 seconds, we assume the program is dead or our connection
to the program has been lost (however this is not necessarily
true). The indicator light will now display this connection
status if applicable (i.e. we are not simulating that program).