[admin] Add "merging code into main" section of workflow wiki (#525)
resolves #490, resolves #491

This change adds the assumptions, details, and example sections for how to merge code into main to the workflow wiki page. This change also includes appendices on how to format commit messages and who should review pull requests.

This change was tested by manual inspection of the rendered content on my fork to verify the GitHub flavored markdown rendered properly.