Use 1 radio repeater LCM instead of 2
resolves #539

We had 2 radio repeater LCM messages of the same structure.
This change deprecates one and renames the other so we can
use the same one for both purposes (initializing the drop
and signalling the completion of the drop).

This change was tested by checking the interface between
nav and the nav simulator still works properly for
dropping the repeater. These are currently the only two
programs that use this code.