[nav-sim] Implement ZED Gimbal (#537)
resolves #471

This changes adds the LCM messages and channels needed for the ZED gimbal.
This change implements the simulation of the ZED gimbal in the navigation simulator and adds hotkeys for manually controlling the simulated gimbal in order to allow for richer testing.

This changes was testing by sending faked ZED gimbal command messages using the LCM send tool. Testing was done to ensure that all the affected services still built. No further integration testing was required since this was the first change related to the ZED gimbal.