[nav] Update Gate Traversal Algorithm (#575)
resolves #340

Replace GateShimmy code with code to drive the rover through the gate by having it drive to a post, then drive through the gate.

This PR was originally on main. However, we have closed the original review and moved it to the auton integration testing branch so we can run it before merging it into the main repository.

These changes were tested by using the navigation simulator to mock several different rover scenarios when searching for a waypoint and turning/driving through the gate posts in the correct direction. This includes if the rover passed through the gate in the incorrect direction (and then reorient itself to pass through the correct direction).