[nav] Added configurable obstacle distance threshold function (#615)
Resolves #499

Added a function (isObstacleInThreshold) to check if an obstacle’s distance from the rover is less than or equal to the distance threshold defined in the nav config file. This function runs whenever we check if an obstacle is detected, so the rover only enters obstacle avoidance state when it is within the distance threshold.

This change was tested in the navigation simulator to make sure obstacle avoidance worked as usual and was only entered when the rover was within a specific distance threshold.